+++ src/io/iobestiary.cpp
@@ -27,7 +27,7 @@
 #include "creatures/players/player.h"
 
 
-bool IOBestiary::parseCharmCombat(Charm* charm, Player* player, Creature* target, int32_t realDamage)
+bool IOBestiary::parseCharmCombat(Charm* charm, Player* player, Creature* target, int64_t realDamage)
 {
 	if (!charm || !player || !target) {
 		return false;
@@ -42,8 +42,8 @@ bool IOBestiary::parseCharmCombat(Charm* charm, Player* player, Creature* target
 			target->addCondition(cripple);
 			player->sendCancelMessage(charm->cancelMsg);
 			return false;
-      }
-		int32_t maxHealth = target->getMaxHealth();
+		}
+		int64_t maxHealth = target->getMaxHealth();
 		charmDamage.primary.type = charm->dmgtype;
 		charmDamage.primary.value = ((-maxHealth * (charm->percent)) / 100);
 		charmDamage.extension = true;
