+++ src/lua/functions/lua_functions_loader.cpp
@@ -268,9 +268,9 @@ void LuaFunctionsLoader::setCreatureMetatable(lua_State* L, int32_t index, const
 
 CombatDamage LuaFunctionsLoader::getCombatDamage(lua_State* L) {
 	CombatDamage damage;
-	damage.primary.value = getNumber<int32_t>(L, -4);
+	damage.primary.value = getNumber<int64_t>(L, -4);
 	damage.primary.type = getNumber<CombatType_t>(L, -3);
-	damage.secondary.value = getNumber<int32_t>(L, -2);
+	damage.secondary.value = getNumber<int64_t>(L, -2);
 	damage.secondary.type = getNumber<CombatType_t>(L, -1);
 
 	lua_pop(L, 4);
